Fixing CSRF (hopefully for good)
It was breaking before because the cookie's Path was set to /api. This was sensible in that only /api routes need CSRF validation, but it also meant that the server would keep resetting the csrf_base token on any request not to /api. It would work in single tab, but broke if the user opened multiple browser tabs and tried to make POST requests that required CSRF validation.

This makes the cookie path to just / and renames csrf_base to csrf_base2 so we don't conflict with users who have the previous cookie already set.