Bringing back reactions in author mode
In #423, I meant to hide the reaction buttons when the author is viewing their own entry, but I accidentally hid the reaction buttons *and* the reactions themselves.

This change brings back both the reaction buttons and the reaction history until I find a cleaner way to hide the buttons in author mode.