Check if authentication is required after checking login state (#518)
* Adding an integration test for expired login

* Check if authentication is required after checking login state

If the What Got Done's store thinks the user is logged in, but UserKit says they're not, they'll end up in a weird state where they stay on a page they're not authenticated to interact with. This adds a check after the UserKit authentication to redirect the user to login if they need authentication on the given page.