Update Google Analytics data automatically
Previously, we relied on an external process to call /api/tasks/refreshGoogleAnalytics periodically. This was a naive approach and required us to refresh analytics and do a lot of datastore writes even when nobody was viewing the site for hours.

This implementation refreshes stats if someone requests pageview stats and it has been more than 3 minutes since the last refresh. Note that it updates it for the next request; the current request will get the stale data, but that's okay, as we assume pageview stats don't change *that* much in a few minutes.